Separate schedule and on-calls react views
