+++ src/views/schedules/show.html.j2
@@ -4,6 +4,6 @@
 {% block content %}
 {# {{ schedule | dump }} #}
 
-<div class="schedule_root" data-id="{{ schedule.scheduleId }}"></div>
+<div class="on_call_root" data-id="{{ schedule.scheduleId }}"></div>
 
 {% endblock %}
